project_dubbing/requirements.txt
ffmpeg-python
pydub
openai-whisper
googletrans==4.0.0-rc1
gTTS
